refactor(web): 리뷰반영
